+++ index.js
@@ -1,1 +1,6 @@
-// empty file for first npm publish
+import MatomoTracker from './src/MatomoTracker';
+import MatomoProvider from './src/MatomoProvider';
+import useMatomo from './src/useMatomo';
+
+export default MatomoTracker;
+export { MatomoProvider, useMatomo };