+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [v0.2.0]
+
+First version to use Matomo tracking
+
+### Added
+
+- created `MatomoTracker` class, `MatomoProvider` context and `useMatomo` hook to provide
+  functionality for tracking with a Matomo instance
+
 ## [v0.1.0]
 
 First npm publish